♻️ Move buttons
